+++ ServiceControlManager/TuningSourceWidget.cpp
@@ -122,7 +122,7 @@ TuningSourceWidget::TuningSourceWidget(quint64 id, QString equipmentId, QString 
 	// Source info
 	//
 	m_infoTable = new QTableView(this);
-	m_infoModel = new QStandardItemModel(staticPropertiesFieldList.count(), 2, this);
+	m_infoModel = new QStandardItemModel(static_cast<int>(staticPropertiesFieldList.count()), 2, this);
 
 	for (int i = 0; i < staticPropertiesFieldList.count(); i++)
 	{
@@ -136,7 +136,7 @@ TuningSourceWidget::TuningSourceWidget(quint64 id, QString equipmentId, QString 
 	// Source state
 	//
 	m_stateTable = new QTableView(this);
-	m_stateModel = new QStandardItemModel(dynamicPropertiesFieldList.count(), 2, this);
+	m_stateModel = new QStandardItemModel(static_cast<int>(dynamicPropertiesFieldList.count()), 2, this);
 
 	for (int i = 0; i < dynamicPropertiesFieldList.count(); i++)
 	{
